+++ src/main.rs
@@ -3,6 +3,11 @@ use juan::data::Event;
 use std::fs;
 use std::io::{self, Read, Write};
 use structopt::StructOpt;
+use std::num::ParseIntError;
+
+fn parse_hex(input: &str) -> Result<u64, ParseIntError> {
+    u64::from_str_radix(input, 16)
+}
 
 #[derive(StructOpt, Debug)]
 struct Opt {
@@ -22,6 +27,10 @@ enum Juan {
     Add,
     Init,
     List,
+    Finish {
+        #[structopt(parse(try_from_str = parse_hex))]
+        id: u64,
+    },
     // Fetch {
     //     #[structopt(long)]
     //     dry_run: bool,
@@ -44,13 +53,16 @@ fn main() {
     match opt.cmd {
         Juan::Add => {
             add_com();
-        },
+        }
         Juan::Init => {
             init_com();
-        },
+        }
         Juan::List => {
             list_com();
-        },
+        }
+        Juan::Finish { id } => {
+            finish_com(id);
+        }
         _ => println!("ELSE!"),
     }
     // println!("{:?}", opt);
@@ -74,6 +86,7 @@ fn write_data(data_set: Vec<Event>) {
     let mut file = fs::OpenOptions::new()
         .write(true)
         .create(true)
+        .truncate(true)
         .open("data.yaml")
         .unwrap();
 
@@ -88,14 +101,16 @@ fn add_com() {
     print!("time (today): ");
     io::stdout().flush().unwrap();
     io::stdin().read_line(&mut time).unwrap();
+    time.pop();
 
     let mut title = String::new();
     print!("title: ");
     io::stdout().flush().unwrap();
     io::stdin().read_line(&mut title).unwrap();
+    title.pop();
 
     let event = Event::from_str(&time, &title);
-    println!("{}", event);
+    println!("One event added:\n{}", event);
     data_set.push(event);
 
     write_data(data_set);
@@ -113,9 +128,20 @@ fn init_com() {
 }
 
 fn list_com() {
-    let mut data_set: Vec<Event> = get_data();
+    let data_set: Vec<Event> = get_data();
 
+    println!("You have {} event(s) in total:", data_set.len());
     for e in &data_set {
         println!("{}", e);
     }
+}
+
+fn finish_com(id: u64) {
+    let mut data_set: Vec<Event> = get_data();
+
+    data_set.retain(|e| e.calculate_hash() != id);
+    
+    println!("Congratulations!");
+    println!("{} event(s) left.", data_set.len());
+    write_data(data_set);
 }