+++ changelog.md
@@ -1,7 +1,14 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## [0.0.1] - 2020-11-23
+## [0.1.1] - 2020-11-24
+### Added
+- Timezone output will be local timezone.
+- Add new commands finish.
+- We use hash as `id` for every event.
+- Change the output of commands add & list.
+
+## [0.1.0] - 2020-11-23
 ### Added
 - We now support three commands (add, list, init) !
 - Add the timezone support (local time by default).